Get triple with subject verb object.

Get triple with subject verb object.

Usage

get_triple(
  triple = NULL,
  subject = NULL,
  verb = NULL,
  object = NULL,
  required = TRUE,
  label = NA,
  limit = NULL,
  within_box = c(NA, NA),
  within_distance = c(NA, NA)
)

Arguments

triple

triple

subject

an anonymous variable (for instance, and by default, "?subject") or item (for instance "wd:Q456"))

verb

the property (for instance "wdt:P190")

object

an anonymous variable (for instance, and by default, "?object") or item (for instance "wd:Q456"))

required

whether to require this triple (Boolean)

label

a vector of variables for which to include a label column (defaults to NA)

limit

the max number of items sent back

within_box

if provided, rectangular bounding box for the triple query. Provided as list(southwest=c(long=...,lat=...),northeast=c(long=...,lat=...))

within_distance

if provided, circular bounding box for the triple query. Provided as list(center=c(long=...,lat=...), radius=...), with radius in kilometers. The center can also be provided as a variable (for instance, "?location") for the center coordinates to be retrieved directly from the query.

Examples

get_triple(s="wd:Q456",v="wdt:P625",o="?coords")
#> # A tibble: 1 × 1
#>   coords                         
#>   <chr>                          
#> 1 Point(4.841388888 45.758888888)
get_triple(t="wd:Q456 wdt:P625 ?coords")
#> # A tibble: 1 × 1
#>   coords                         
#>   <chr>                          
#> 1 Point(4.841388888 45.758888888)