Send SPARQL query to endpoint and get tibble as a result

Send SPARQL query to endpoint and get tibble as a result

Usage

send_sparql(
  .query,
  endpoint = "Wikidata",
  user_agent = getOption("glitter.ua",
    "glitter R package (https://github.com/lvaudor/glitter)"),
  max_tries = getOption("glitter.max_tries", 3L),
  max_seconds = getOption("glitter.max_seconds", 120L),
  timeout = getOption("glitter.timeout", 1000L),
  request_type = c("url", "body-form"),
  dry_run = FALSE
)

Arguments

.query

a string corresponding to a SPARQL query

endpoint

a string or url corresponding to a SPARQL endpoint. Defaults to "Wikidata"

user_agent

a string indicating the user agent to send with the query.

max_tries, max_seconds

Cap the maximum number of attempts with max_tries or the total elapsed time from the first request with max_seconds. If neither option is supplied (the default), req_perform() will not retry.

timeout

maximum number of seconds to wait (httr2::req_timeout()).

request_type

a string indicating how the query should be sent: in the URL (url, default, most common) or as a body form (body-form).

dry_run

Boolean indicating whether to return the output of httr2::req_dry_run() rather than of httr2::req_perform. Useful for debugging.

Examples

metro_query='SELECT ?item ?itemLabel ?coords
{
 ?item wdt:P361 wd:Q1552;
 wdt:P625 ?coords.
 OPTIONAL{?item wdt:P1619 ?date.}
 SERVICE wikibase:label { bd:serviceParam wikibase:language "en" . }
} ORDER BY ?itemLabel
'
send_sparql(metro_query)
#> # A tibble: 0 × 3
#> # ℹ 3 variables: item <chr>, itemLabel <chr>, coords <chr>