Add a triple pattern statement to a query

Add a triple pattern statement to a query

Usage

spq_add(
  .query = NULL,
  .triple_pattern = NULL,
  .subject = NULL,
  .verb = NULL,
  .object = NULL,
  .prefixes = NULL,
  .required = TRUE,
  .label = NA,
  .within_box = c(NA, NA),
  .within_distance = c(NA, NA)
)

Arguments

.query

query

.triple_pattern

the triple pattern statement (replaces arguments subject verb and object)

.subject

an anonymous variable (for instance, and by default, "?subject") or item (for instance "wd:Q456"))

.verb

the property (for instance "wdt:P190")

.object

an anonymous variable (for instance, and by default, "?object") or item (for instance "wd:Q456"))

.prefixes

Custom prefixes

.required

whether the existence of a value for the triple is required or not (defaults to TRUE). If set to FALSE, then other triples in the query are returned even if this particular triple is missing)

.label

a vector of variables for which to include a label column (defaults to NA)

.within_box

if provided, rectangular bounding box for the triple query. Provided as list(southwest=c(long=...,lat=...),northeast=c(long=...,lat=...))

.within_distance

if provided, circular bounding box for the triple query. Provided as list(center=c(long=...,lat=...), radius=...), with radius in kilometers. The center can also be provided as a variable (for instance, "?location") for the center coordinates to be retrieved directly from the query.

Details

The arguments .subject, .verb, .object are most useful for programmatic usage, they are actually used within glitter code itself.

Examples

# find the cities
spq_init() %>%
spq_add("?city wdt:P31/wdt:P279* wd:Q515", .label="?city") %>%
# and their populations
spq_add("?city wdt:P1082 ?pop", .required = FALSE) %>%
# in a bounding box
spq_add("?city wdt:P625 ?coords", .within_box = list(southwest = c(3,43), northeast = c(7,47))) %>%
# limit to 10 lines
spq_head(n = 10)
#> 
#> SELECT ?city ?cityLabel ?pop ?coords
#> WHERE{
#> 
#> ?city wdt:P31/wdt:P279* wd:Q515.
#> OPTIONAL {?city wdt:P1082 ?pop.}
#> SERVICE wikibase:box {
#> ?city wdt:P625 ?coords.
#> bd:serviceParam wikibase:cornerSouthWest 'Point(3 43)'^^geo:wktLiteral.
#> bd:serviceParam wikibase:cornerNorthEast 'Point(7 47)'^^geo:wktLiteral.
#> }
#> 
#> SERVICE wikibase:label { bd:serviceParam wikibase:language "en".}
#> }
#> 
#> LIMIT 10

if (FALSE) {
# find the individuals of the species
spq_init() %>%
spq_add("?mayor wdt:P31 ?species") %>%
# dog, cat or chicken
spq_add("?species %in% {c('wd:144','wd:146', 'wd:780')}") %>%
# who occupy the function
spq_add("?mayor p:P39 ?node") %>%
# of mayor
spq_add("?node ps:P39 wd:Q30185") %>%
# of some places
spq_add("?node pq:P642 ?place") %>%
spq_perform()
}